// src/app/core/api/generated/schemas/carBrandResponseDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts merged API
 * OpenAPI spec version: 1.0
 */

export interface CarBrandResponseDto {
  car_brand_name: string;
  brand_image: string;
  brand_url: string;
  is_active: boolean;
  is_deleted: boolean;
  car_brand_id: number;
}