// src/app/core/api/generated/schemas/productCategoryAttributeControllerWorkspaceParams.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * stallio-auto-parts API
 * OpenAPI spec version: 1.0
 */

export type ProductCategoryAttributeControllerWorkspaceParams = {
  page?: number;
  limit?: number;
  search?: string;
  category_search?: string;
  brand_search?: string;
  model_search?: string;
  product_search?: string;
  selector_limit?: number;
  category_id?: number;
  car_brand_id?: number;
  model_id?: number;
  product_id?: number;
};